+++ packages/BiostructureViewer/src/package-test.ts
@@ -6,6 +6,7 @@ import {runTests, tests, TestContext} from '@datagrok-libraries/utils/src/test';
 
 import './tests/pdb-helper-tests';
 import './tests/pdb-tests';
+import './tests/ngl-gl-service';
 import './tests/pdb-grid-cell-renderer-tests';
 import './tests/molstar-preview-tests';
 import './tests/molstar-viewer-tests';