+++ packages/BiostructureViewer/src/viewers/saguaro-viewer.ts
@@ -133,8 +133,7 @@ export class SaguaroViewer extends DG.JsViewer {
 
   override detach(): void {
     const superDetach = super.detach.bind(this);
-    this.detachPromise = this.detachPromise.then(async () => {
-      await this.viewPromise;
+    this.viewPromise = this.viewPromise.then(async () => {
       if (this.setDataInProgress) return; // check setDataInProgress synced
       if (this.viewed) {
         await this.destroyView('detach'); // detach
@@ -157,8 +156,6 @@ export class SaguaroViewer extends DG.JsViewer {
           await this.destroyView('setData'); // setData
           this.viewed = false;
         }
-
-        await this.detachPromise;
         // TODO: Data
 
         if (!this.viewed) {
@@ -175,7 +172,6 @@ export class SaguaroViewer extends DG.JsViewer {
 
   // -- View --
   private viewPromise: Promise<void> = Promise.resolve();
-  private detachPromise: Promise<void> = Promise.resolve();
   private setDataInProgress: boolean = false;
 
   private viewerDivId: string;